requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
sqlalchemy==1.3.23
python-dotenv==0.15.0
pydantic==1.8.1
